+++ packages/d1/src/components/Ui/Avatar/D1Avatar.stories.ts
@@ -31,4 +31,29 @@ export const Avatar: Story = {
 }
 
 // :story-items [!] System label / Системная метка
+export const AvatarBasic: Story = {
+  name: 'Базовые',
+  render: () => ({
+    components: { D1Avatar },
+    template: `
+        <div class="wiki-storybook-flex">
+          <D1Avatar label="John Doe"/>
+          <D1Avatar src="https://images.unsplash.com/photo-1534528741775-53994a69daeb?auto=format&fit=crop&w=100&q=80" label="User Image"/>
+          <D1Avatar label="Online User" badgeDot/>
+        </div>
+    `
+  })
+}
+export const AvatarReduction: Story = {
+  name: 'Сокращение (reduction)',
+  render: () => ({
+    components: { D1Avatar },
+    template: `
+        <div class="wiki-storybook-flex">
+          <D1Avatar label="Иван Иванович" :reduction="true"/>
+          <D1Avatar label="Иван Иванович" :reduction="false"/>
+        </div>
+    `
+  })
+}
 // :story-items [!] System label / Системная метка