+++ packages/d1/src/components/Ui/Avatar/wikiData.ts
@@ -5,6 +5,26 @@ import { defaults } from './props'
 
 const propsNames: StorybookProps = [
   // :propsList [!] System label / Системная метка
+  { name: 'ariaLabel', type: 'string' },
+  { name: 'badge', type: 'string | number | ConstrBind<BadgePropsBasic>' },
+  { name: 'badgeAttrs', type: 'ConstrBind<BadgePropsBasic>' },
+  { name: 'badgeDot', type: 'boolean' },
+  { name: 'detail', type: 'Record<string, any>' },
+  { name: 'disabled', type: 'boolean' },
+  { name: 'focus', type: 'boolean' },
+  { name: 'href', type: 'string' },
+  { name: 'image', type: 'string | ConstrBind<ImagePropsBasic>' },
+  { name: 'imageAttrs', type: 'ConstrBind<ImagePropsBasic>' },
+  { name: 'isSkeleton', type: 'boolean' },
+  { name: 'label', type: 'NumberOrString' },
+  { name: 'labelId', type: 'string' },
+  { name: 'palette', type: 'string', option: ['red', 'orange', 'amber', 'yellow', 'lime', 'green', 'emerald', 'teal', 'cyan', 'sky', 'blue', 'indigo', 'violet', 'purple', 'fuchsia', 'pink', 'rose', 'slate', 'gray', 'zinc', 'neutral', 'stone', 'black', 'white'] },
+  { name: 'primary', type: 'boolean' },
+  { name: 'readonly', type: 'boolean' },
+  { name: 'reduction', type: 'boolean' },
+  { name: 'size', type: 'string', option: ['xs', 'sm', 'md', 'lg', 'xl'] },
+  { name: 'to', type: 'string | RouteLocationAsRelativeGeneric | RouteLocationAsPathGeneric' },
+  { name: 'value', type: 'EventClickValue[\'value\']' }
   // :propsList [!] System label / Системная метка
 ]